+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- exit with non-zero exit code if any violations are found
+
 ## [0.0.2] - 2026-02-15
 
 - MVP: Basic architecture enforcement with layered architecture support